Crate bevy_platform

Platform compatibility support for first-party Bevy engine crates.

Modules

cell

Provides cell primitives.

cfg

Provides helpful configuration macros, allowing detection of platform features such as alloc or std without explicit features.

collections

Provides HashMap and HashSet from hashbrown with some customized defaults.

hash

Provides replacements for std::hash items using foldhash.

prelude

Frequently used items which would typically be included in most contexts.

sync

Provides various synchronization alternatives to language primitives.

thread

Provides sleep for all platforms.

time

Provides Instant for all platforms.