bevy_platform

Module collections

Provides HashMap and HashSet from hashbrown with some customized defaults.

Also provides the HashTable type, which is specific to hashbrown.

Re-exports

pub use hash_map::HashMap;

pub use hash_set::HashSet;

Modules

hash_map

Provides HashMap based on hashbrown’s implementation. Unlike hashbrown::HashMap, HashMap defaults to FixedHasher instead of RandomState. This provides determinism by default with an acceptable compromise to denial of service resistance in the context of a game engine.

hash_set

Provides HashSet based on hashbrown’s implementation. Unlike hashbrown::HashSet, HashSet defaults to FixedHasher instead of RandomState. This provides determinism by default with an acceptable compromise to denial of service resistance in the context of a game engine.

hash_table

Provides HashTable

Structs

HashTable

Low-level hash table with explicit hashing.

Traits

Equivalent

Key equivalence trait.