bevy_platform::collections

Module hash_set

Provides HashSet based on hashbrown’s implementation. Unlike hashbrown::HashSet, HashSet defaults to FixedHasher instead of RandomState. This provides determinism by default with an acceptable compromise to denial of service resistance in the context of a game engine.

Structs

Difference

A lazy iterator producing elements in the difference of HashSets.

Drain

A draining iterator over the items of a HashSet.

ExtractIf

A draining iterator over entries of a HashSet which don’t satisfy the predicate f.

HashSet

New-type for HashSet with FixedHasher as the default hashing provider. Can be trivially converted to and from a hashbrown HashSet using From.

Intersection

A lazy iterator producing elements in the intersection of HashSets.

IntoIter

An owning iterator over the items of a HashSet.

Iter

An iterator over the items of a HashSet.

OccupiedEntry

A view into an occupied entry in a HashSet. It is part of the Entry enum.

SymmetricDifference

A lazy iterator producing elements in the symmetric difference of HashSets.

Union

A lazy iterator producing elements in the union of HashSets.

VacantEntry

A view into a vacant entry in a HashSet. It is part of the Entry enum.

Type Aliases

Entry

Shortcut for Entry with FixedHasher as the default hashing provider.